# fixtures/task-solutions/nullable-patch-codegen/internal/repository/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package generated

import (
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type Task struct {
	ID        string
	Title     string
	CreatedAt time.Time
	DueAt     pgtype.Timestamptz
}