# fixtures/task-solutions/optimistic-locking-codegen/internal/repository/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package generated

import (
	"time"
)

type Task struct {
	ID        string
	Title     string
	CreatedAt time.Time
	Version   int64
}